Added a rank score system for cases where we get multiple function vals.
Sometimes there is more than one functional annotation added to a
variant call.  In this case, use a heuristic ranking approach to select
the most impactful one and only report that one.  In most cases, we see
two missense annotations, and so there's really no ranking needed.  But,
if there is a nonsense vs a missesne, I'd rather pick that one out.