Fixed formatting warning / bug.
Some newer versions of perl complain about a ``Redundant argument in
printf``, which is caused by the extra TVC / LIA tag in each entry.
Popping that value off the array before printing now.  Also, change the
``NULL`` padding string to be for any member of the array and not just
indices 9 - 13; not sure why it was that way in the first place anymore.