Fixed formatted print of data with NOCALLs.
The new pop of the `tvc` or `lia` flag will not work for data when we
don't have OVAT annotation output since the length of the data array is
different.  Handling that output array a little better now so that we
can switch between the two different length arrays, depending on the
collected data.