Added get_gene_info.R, a wrapper around the biomaRt R package.
Had this script fragment laying around for a while, and finally made
some helpful refinements such that this is ready to be of use. If you
know the ensemble_transcript_id, refseq_mrna, entrezgene_id,
hgnc_symbol, or ensembl_gene_id, search the data base and return those
terms. Helpful for cross-mapping IDs.